@@ -1,45 +1,43 @@
 {
-    "name": "default",
-    "version": "0.0.0",
-    "license": "MIT",
-    "scripts": {
-        "ng": "ng",
-        "start": "ng serve",
-        "build": "ng build --prod",
-        "test": "ng test",
-        "lint": "ng lint",
-        "e2e": "ng e2e",
-        "postinstall": "npm run tsformat",
-        "tsformat": "tsfmt -r"
-    },
-    "private": true,
-    "dependencies": {
-        "@angular/animations": "^5.2.0",
-        "@angular/common": "^5.2.0",
-        "@angular/compiler": "^5.2.0",
-        "@angular/core": "^5.2.0",
-        "@angular/forms": "^5.2.0",
-        "@angular/http": "^5.2.0",
-        "@angular/platform-browser": "^5.2.0",
-        "@angular/platform-browser-dynamic": "^5.2.0",
-        "@angular/router": "^5.2.0",
-        "core-js": "^2.4.1",
-        "rxjs": "^5.5.6",
-        "zone.js": "^0.8.19",
-        "jquery": "^3.2.1",
-        "@types/jquery": "^3.2.6",
-        "enhanced-resolve": "^3.3.0",
-        "classlist.js": "^1.1.20150312",
-        "@ng-bootstrap/ng-bootstrap": "^1.0.0-beta.3",
-        "primeng": "^4.1.3",
-        "ng-recaptcha": "^3.0.0",
-        "typescript-formatter": "^5.2.0"
-    },
-    "devDependencies": {
-        "@angular-devkit/build-angular": "~0.8.0",
-        "@angular/cli": "^6.2.3",
-        "@angular/compiler-cli": "^5.2.0",
-        "@angular/language-service": "^5.2.0",
-        "typescript": "~2.5.3"
-    }
+  "name": "default",
+  "version": "0.0.0",
+  "license": "MIT",
+  "scripts": {
+    "ng": "ng",
+    "start": "ng serve",
+    "build": "ng build --prod",
+    "test": "ng test",
+    "lint": "ng lint",
+    "e2e": "ng e2e"
+  },
+  "private": true,
+  "dependencies": {
+    "@angular/animations": "^7.0.1",
+    "@angular/common": "^7.0.1",
+    "@angular/compiler": "^7.0.1",
+    "@angular/core": "^7.0.1",
+    "@angular/forms": "^7.0.1",
+    "@angular/http": "^7.0.1",
+    "@angular/platform-browser": "^7.0.1",
+    "@angular/platform-browser-dynamic": "^7.0.1",
+    "@angular/router": "^7.0.1",
+    "@ng-bootstrap/ng-bootstrap": "^4.0.0",
+    "@types/jquery": "^3.2.6",
+    "classlist.js": "^1.1.20150312",
+    "core-js": "^2.4.1",
+    "enhanced-resolve": "^3.3.0",
+    "jquery": "^3.2.1",
+    "ng-recaptcha": "^4.2.1",
+    "primeng": "^4.1.3",
+    "rxjs": "^6.3.3",
+    "typescript-formatter": "^5.2.0",
+    "zone.js": "^0.8.19"
+  },
+  "devDependencies": {
+    "@angular-devkit/build-angular": "~0.10.0",
+    "@angular/cli": "^7.0.3",
+    "@angular/compiler-cli": "^7.0.1",
+    "@angular/language-service": "^7.0.1",
+    "typescript": "~3.1.4"
+  }
 }
