@@ -1,7 +1,7 @@
 import { Injectable } from "@angular/core";
 import { ActivatedRouteSnapshot, CanActivate, Router, RouterStateSnapshot } from "@angular/router";
 import { UserService } from "../_services/user.service";
-import { Observable } from "rxjs/Rx";
+import { Observable } from "rxjs";
 
 @Injectable()
 export class AuthGuard implements CanActivate {
@@ -10,21 +10,15 @@ export class AuthGuard implements CanActivate {
     }
 
     canActivate(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): Observable<boolean> | boolean {
-        let currentUser = JSON.parse(localStorage.getItem('currentUser'));
-        return this._userService.verify().map(
-            data => {
-                if (data !== null) {
-                    // logged in so return true
-                    return true;
-                }
-                // error when verify so redirect to login page with the return url
-                this._router.navigate(['/login'], { queryParams: { returnUrl: state.url } });
-                return false;
-            },
-            error => {
-                // error when verify so redirect to login page with the return url
-                this._router.navigate(['/login'], { queryParams: { returnUrl: state.url } });
-                return false;
-            });
+        console.log("authguard", localStorage.getItem('currentUser'));
+        let data = localStorage.getItem('currentUser');         
+        if (data !== null) {
+            // logged in so return true
+            console.log("logged in", data);
+            return true;
+        }
+        // error when verify so redirect to login page with the return url
+        this._router.navigate(['/login'], { queryParams: { returnUrl: state.url } });
+        return false;
     }
 }
