@@ -1,23 +1,16 @@
 import { Injectable } from "@angular/core";
-import { Http, Response } from "@angular/http";
-import "rxjs/add/operator/map";
+import { HttpClient   } from "@angular/common/http";
+import { map } from 'rxjs/operators';
 
 @Injectable()
 export class AuthenticationService {
 
-    constructor(private http: Http) {
+    constructor(private http: HttpClient  ) {
     }
 
     login(email: string, password: string) {
-        return this.http.post('/api/authenticate', JSON.stringify({ email: email, password: password }))
-            .map((response: Response) => {
-                // login successful if there's a jwt token in the response
-                let user = response.json();
-                if (user && user.token) {
-                    // store user details and jwt token in local storage to keep user logged in between page refreshes
-                    localStorage.setItem('currentUser', JSON.stringify(user));
-                }
-            });
+        return  localStorage.setItem('currentUser', JSON.stringify({email, password}));
+             
     }
 
     logout() {
