@@ -1,8 +1,8 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
-import { BaseRequestOptions, HttpModule } from '@angular/http';
-import { MockBackend } from '@angular/http/testing';
+import { HttpClientModule } from '@angular/common/http';
+
 
 import { AuthRoutingModule } from './auth-routing.routing';
 import { AuthComponent } from './auth.component';
@@ -12,7 +12,7 @@ import { AuthGuard } from './_guards/auth.guard';
 import { AlertService } from './_services/alert.service';
 import { AuthenticationService } from './_services/authentication.service';
 import { UserService } from './_services/user.service';
-import { fakeBackendProvider } from './_helpers/index';
+
 
 @NgModule({
     declarations: [
@@ -23,18 +23,14 @@ import { fakeBackendProvider } from './_helpers/index';
     imports: [
         CommonModule,
         FormsModule,
-        HttpModule,
+        HttpClientModule,
         AuthRoutingModule,
     ],
     providers: [
         AuthGuard,
         AlertService,
         AuthenticationService,
-        UserService,
-        // api backend simulation
-        fakeBackendProvider,
-        MockBackend,
-        BaseRequestOptions,
+        UserService       
     ],
     entryComponents: [AlertComponent],
 })
