@@ -59,53 +59,37 @@ export class AuthComponent implements OnInit {
 
     signin() {
         this.loading = true;
-        this._authService.login(this.model.email, this.model.password).subscribe(
-            data => {
-                this._router.navigate([this.returnUrl]);
-            },
-            error => {
-                this.showAlert('alertSignin');
-                this._alertService.error(error);
-                this.loading = false;
-            });
+        this._authService.login(this.model.email, this.model.password)
+        this._router.navigate([this.returnUrl]);
+    
+        this.showAlert('alertSignin');
+        this.loading = false;
+          
     }
 
     signup() {
         this.loading = true;
-        this._userService.create(this.model).subscribe(
-            data => {
-                this.showAlert('alertSignin');
-                this._alertService.success(
-                    'Thank you. To complete your registration please check your email.',
-                    true);
-                this.loading = false;
-                LoginCustom.displaySignInForm();
-                this.model = {};
-            },
-            error => {
-                this.showAlert('alertSignup');
-                this._alertService.error(error);
-                this.loading = false;
-            });
+       
+        this._alertService.success(
+            'Thank you. To complete your registration please check your email.',
+            true);
+        this.loading = false;
+        LoginCustom.displaySignInForm();
+        this.model = {};
+         
     }
 
     forgotPass() {
         this.loading = true;
-        this._userService.forgotPassword(this.model.email).subscribe(
-            data => {
-                this.showAlert('alertSignin');
-                this._alertService.success(
-                    'Cool! Password recovery instruction has been sent to your email.',
-                    true);
-                this.loading = false;
-                LoginCustom.displaySignInForm();
-                this.model = {};
-            },
-            error => {
-                this.showAlert('alertForgotPass');
-                this._alertService.error(error);
-                this.loading = false;
-            });
+        
+        this.showAlert('alertSignin');
+        this._alertService.success(
+            'Cool! Password recovery instruction has been sent to your email.',
+            true);
+        this.loading = false;
+        LoginCustom.displaySignInForm();
+        this.model = {};
+    
     }
 
     showAlert(target) {
